Create address_book.py
I'm pretty happy with this; however, I can't figure out how to ignore the command to create the table if the table already exists. I don't want to just comment it out, because I want somebody to be able to run this program anytime they want to without having to go in and comment out the code to create the table the second time they run the program.